requirements.txt
Flask
Flask-PyMongo
Flask-SQLAlchemy
Flask-Smorest
marshmallow
python-dotenv
Flask-cors